extract ga4gh service messages into dedicated class
