+++ src/main/scala/org/broadinstitute/dsde/agora/server/ga4gh/Ga4ghService.scala
@@ -3,11 +3,10 @@ package org.broadinstitute.dsde.agora.server.ga4gh
 import akka.actor.Props
 import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.agora.server.dataaccess.permissions.PermissionsDataSource
+import org.broadinstitute.dsde.agora.server.ga4gh.Ga4ghServiceMessages._
 import org.broadinstitute.dsde.agora.server.ga4gh.Models._
 import org.broadinstitute.dsde.agora.server.model.AgoraEntityType
 import org.broadinstitute.dsde.agora.server.webservice.PerRequestCreator
-import org.broadinstitute.dsde.agora.server.webservice.handlers.QueryHandler
-import org.broadinstitute.dsde.agora.server.webservice.util.ServiceMessages._
 import spray.http.{MediaTypes, StatusCodes}
 import spray.httpx.SprayJsonSupport._
 import spray.json.DefaultJsonProtocol._
@@ -44,45 +43,37 @@ abstract class Ga4ghService(permissionsDataSource: PermissionsDataSource)
           complete(toolClassesResponse)
         } ~
         path("tools") { requestContext =>
-          // tools endpoint
           // TODO: query params and response headers
           val message = QueryPublicTools(requestContext)
           perRequest(requestContext, queryHandler, message)
         } ~
         path("tools" / Segment) { id => requestContext =>
-          // tools/{id} endpoint
           val entity = entityFromArguments(id)
           val message = QueryPublicTool(requestContext, entity)
           perRequest(requestContext, queryHandler, message)
         } ~
         path("tools" / Segment / "versions") { id => requestContext =>
-          // tools/{id}/versions endpoint
           val entity = entityFromArguments(id)
           val message = QueryPublic(requestContext, entity)
           perRequest(requestContext, queryHandler, message)
         } ~
         path("tools" / Segment / "versions" / Segment) { (id, versionId) => requestContext =>
-          // tools/{id}/versions/{version-id} endpoint
           val entity = entityFromArguments(id, versionId)
           val message = QueryPublicSingle(requestContext, entity)
           perRequest(requestContext, queryHandler, message)
         } ~
         path("tools" / Segment / "versions" / Segment / "dockerfile") { (id, versionId) =>
-          // TODO: tools/{id}/versions/{version-id}/dockerfile endpoint
           complete(spray.http.StatusCodes.NotImplemented)
         } ~
         path("tools" / Segment / "versions" / Segment / Segment / "descriptor") { (id, versionId, descriptorType) => requestContext =>
-          // /tools/{id}/versions/{version-id}/{type}/descriptor endpoint
           val entity = entityFromArguments(id, versionId)
           val message = QueryPublicSinglePayload(requestContext, entity, parseDescriptorType(descriptorType))
           perRequest(requestContext, queryHandler, message)
         } ~
         path("tools" / Segment / "versions" / Segment / Segment / "descriptor" / Segment) { (id, versionId, descriptorType, relativePath) =>
-          // /tools/{id}/versions/{version-id}/{type}/descriptor/{relative-path} endpoint not supported
           complete(spray.http.StatusCodes.NotImplemented)
         } ~
         path("tools" / Segment / "versions" / Segment / Segment / "tests") { (id, versionId, descriptorType) =>
-          // /tools/{id}/versions/{version-id}/{type}/tests endpoint not supported
           complete(spray.http.StatusCodes.NotImplemented)
         }
       }
